feat(stats screen): stats Screen & bugFix

+++ data/src/main/java/com/yestms/driver/android/data/remote/response/loads/LoadResponse.kt
@@ -6,7 +6,7 @@ data class LoadResponse(
     val id: Int?,
     val loadId: String?,
     val activationLink: String?,
-    val rate: Int?,
+    val rate: Long?,
     val mileage: Int?,
     val pickUpNote: String?,
     val pickUpLocation: String?,
