+++ domain/src/main/java/com/yestms/driver/android/domain/model/loads/LoadModel.kt
@@ -4,7 +4,7 @@ data class LoadModel(
     val id: Int,
     val loadId: String,
     val activationLink: String,
-    val rate: Int,
+    val rate: Long,
     val mileage: Int,
     val pickUpNote: String,
     val pickUpLocation: String,
